refactor : test code divison class
