+++ src/test/java/controller/LottoControllerTest.java
@@ -1,0 +1,24 @@
+package controller;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class LottoControllerTest {
+
+    @Test
+    void totalLotto() {
+    }
+
+    @Test
+    void resultReturn() {
+    }
+
+    @Test
+    @DisplayName("당첨 결과 반환")
+    public void winningStatus() {
+        LottoController controller = new LottoController();
+
+    }
+}
